local: implement dynamic logging segregation and refactor telemetry coordinator to m365_telemetry.py
