+++ telemetry/mailbox_usage.py
@@ -27,8 +27,8 @@ from core.graph.directory import DirectoryService
 from core.powershell.client import PowerShellClient
 from core.powershell.mailbox import MailboxStatsService
 
-# Bind to the async logger initialized in license_usage.py
-usage_logger = logging.getLogger("LicenseUsageAsyncLogger")
+# Bind to the async logger initialized in m365_telemetry.py
+usage_logger = logging.getLogger("M365TelemetryAsyncLogger")
 
 # =================================================================================
 # CONSTANTS & STYLES (Imported from shared styles)
